refactor(quartzite-macros): move make_expand! macro from lib.rs to util.rs
