+++ quartzite-macros/src/lib.rs
@@ -11,19 +11,8 @@
 
 use proc_macro::TokenStream;
 
-macro_rules! make_expand {
-    () => {
-        pub(crate) fn expand(input: proc_macro2::TokenStream) -> proc_macro2::TokenStream {
-            match parse::parse(input) {
-                Ok(ir) => codegen::codegen(&ir),
-                Err(e) => e.to_compile_error(),
-            }
-        }
-    };
-}
-pub(crate) use make_expand;
-
 mod util;
+pub(crate) use util::make_expand;
 
 mod extend;
 mod meta_enum;
