+++ quartzite-macros/src/util.rs
@@ -158,6 +158,18 @@ pub(crate) fn crate_root_from(
     }
 }
 
+macro_rules! make_expand {
+    () => {
+        pub(crate) fn expand(input: proc_macro2::TokenStream) -> proc_macro2::TokenStream {
+            match parse::parse(input) {
+                Ok(ir) => codegen::codegen(&ir),
+                Err(e) => e.to_compile_error(),
+            }
+        }
+    };
+}
+pub(crate) use make_expand;
+
 #[cfg(test)]
 mod tests {
     use super::*;
